feat: add GrammarSidebar component with smooth scrolling and mobile menu toggle
- Implemented GrammarSidebar component to provide a table of contents for grammar sections.
- Added Intersection Observer to highlight the active section based on scroll position.
- Included smooth scrolling functionality for anchor links.
- Created a mobile-friendly toggle for the sidebar menu.
- Styled the sidebar to be sticky and responsive.
